+++ package.json
@@ -44,7 +44,6 @@
   "dependencies": {
     "classnames": "^2.2.6",
     "express": "^4.16.3",
-    "handlebars": "^4.0.11",
     "highlight.js": "^9.12.0",
     "i18next": "^11.8.0",
     "i18next-express-middleware": "^1.3.2",
@@ -73,7 +72,6 @@
   "devDependencies": {
     "@types/classnames": "^2.2.4",
     "@types/express": "^4.16.0",
-    "@types/handlebars": "^4.0.39",
     "@types/highlight.js": "^9.12.3",
     "@types/i18next": "^8.4.5",
     "@types/i18next-express-middleware": "^0.0.33",
