+++ src/public/index.html
@@ -5,17 +5,11 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>{{ title }}</title>
-  {{{ css }}}
-  <script>
-    window.__APP_INITIAL_TUTURE__ = {{ tuture }};
-    window.__APP_INITIAL_DIFF__ = {{ diff }};
-    window.__APP_INITIAL_I18N__ = {{ i18n }};
-  </script>
+  <title>Tuture</title>
 </head>
 
 <body>
-  <div id="root">{{{ body }}}</div>
+  <div id="root"></div>
 </body>
 
 </html>
