+++ webpack.dev.js
@@ -67,7 +67,7 @@ const serverConfig = merge(base, {
     __dirname: false,
     __filename: false,
   },
-  entry: './src/index.tsx',
+  entry: './src/index.ts',
   output: {
     filename: 'server.js',
     pathinfo: false,
